ref: Move `httpRequestToRequestData` to core, avoiding HTTP dependency
