+++ dev-packages/e2e-tests/test-applications/node-exports-test-app/scripts/consistentExports.ts
@@ -21,7 +21,6 @@ const NODE_EXPORTS_IGNORE = [
   'SentryContextManager',
   'validateOpenTelemetrySetup',
   'preloadOpenTelemetry',
-  'httpRequestToRequestEventData',
 ];
 
 const nodeExports = Object.keys(SentryNode).filter(e => !NODE_EXPORTS_IGNORE.includes(e));
