+++ packages/google-cloud-serverless/src/gcpfunction/http.ts
@@ -2,19 +2,13 @@ import {
   SEMANTIC_ATTRIBUTE_SENTRY_ORIGIN,
   SEMANTIC_ATTRIBUTE_SENTRY_SOURCE,
   handleCallbackErrors,
+  httpRequestToRequestData,
   isString,
   logger,
   setHttpStatus,
   stripUrlQueryAndFragment,
 } from '@sentry/core';
-import {
-  captureException,
-  continueTrace,
-  flush,
-  getCurrentScope,
-  httpRequestToRequestEventData,
-  startSpanManual,
-} from '@sentry/node';
+import { captureException, continueTrace, flush, getCurrentScope, startSpanManual } from '@sentry/node';
 import { DEBUG_BUILD } from '../debug-build';
 import { domainify, markEventUnhandled, proxyFunction } from '../utils';
 import type { HttpFunction, WrapperOptions } from './general';
@@ -52,7 +46,7 @@ function _wrapHttpFunction(fn: HttpFunction, options: Partial<WrapperOptions>): 
     const baggage = req.headers?.baggage;
 
     return continueTrace({ sentryTrace, baggage }, () => {
-      const normalizedRequest = httpRequestToRequestEventData(req);
+      const normalizedRequest = httpRequestToRequestData(req);
       getCurrentScope().setSDKProcessingMetadata({ normalizedRequest });
 
       return startSpanManual(
