+++ packages/nextjs/src/common/pages-router-instrumentation/wrapApiHandlerWithSentry.ts
@@ -3,6 +3,7 @@ import {
   SEMANTIC_ATTRIBUTE_SENTRY_SOURCE,
   captureException,
   continueTrace,
+  httpRequestToRequestData,
   setHttpStatus,
   startSpanManual,
   withIsolationScope,
@@ -10,7 +11,6 @@ import {
 import { isString, logger, objectify } from '@sentry/core';
 
 import { vercelWaitUntil } from '@sentry/core';
-import { httpRequestToRequestEventData } from '@sentry/node';
 import type { NextApiRequest } from 'next';
 import type { AugmentedNextApiResponse, NextApiHandler } from '../types';
 import { flushSafelyWithTimeout } from '../utils/responseEnd';
@@ -66,7 +66,7 @@ export function wrapApiHandlerWithSentry(apiHandler: NextApiHandler, parameteriz
             },
             () => {
               const reqMethod = `${(req.method || 'GET').toUpperCase()} `;
-              const normalizedRequest = httpRequestToRequestEventData(req);
+              const normalizedRequest = httpRequestToRequestData(req);
 
               isolationScope.setSDKProcessingMetadata({ normalizedRequest });
               isolationScope.setTransactionName(`${reqMethod}${parameterizedRoute}`);
