+++ packages/node/src/index.ts
@@ -1,5 +1,4 @@
 export { httpIntegration } from './integrations/http';
-export { httpRequestToRequestEventData } from './integrations/http/SentryHttpInstrumentation';
 export { nativeNodeFetchIntegration } from './integrations/node-fetch';
 export { fsIntegration } from './integrations/fs';
